+++ R/github_download_files.R
@@ -7,10 +7,13 @@
 github_download_files <- function(filelist,
                                   download_dir = tempdir(),
                                   overwrite = FALSE,
+                                  timeout = 5*60,
                                   nThread = 1,
                                   verbose = TRUE) {
     messager("+ Downloading", length(filelist), "files...", v = verbose)
-    local_files <- unlist(parallel::mclapply(filelist, function(x) {
+    local_files <- parallel::mclapply(stats::setNames(filelist,
+                                                      filelist), 
+                                             function(x) {
         message_parallel(paste("Downloading:", x))
         branch <- stringr::str_split(string = x, pattern = "/")[[1]][7]
         folder_structure <- paste(stringr::str_split(
@@ -26,10 +29,11 @@ github_download_files <- function(filelist,
             recursive = TRUE
         )
         if (!file.exists(destfile) & overwrite == FALSE) {
-            options(timeout = 15 * 60)
-            utils::download.file(url = x, destfile = destfile)
+            options(timeout = timeout)
+            utils::download.file(url = x, 
+                                 destfile = destfile)
         }
         return(destfile)
-    }, mc.cores = nThread))
+    }, mc.cores = nThread)
     return(local_files)
 }