+++ NEWS.md
@@ -11,6 +11,7 @@
     - `get_SNPgroup_counts`
     - `results_report`
     - `merge_finemapping_results`
+* `portal_query`: add arg to return results in table format.
 
 ## Bug fixes